Support opening files bigger than 2 GiB
The real technical limit is with lines bigger than 2 GiB and buffers
with more than 2 Gi lines, refactor buffer loading to make it possible
to load those files.

Fix an overflow with the hash_data function at the same time

+++ src/hash.cc
@@ -35,10 +35,10 @@ size_t hash_data(const char* input, size_t len)
     constexpr uint32_t c1 = 0xcc9e2d51;
     constexpr uint32_t c2 = 0x1b873593;
 
-    const int nblocks = len / 4;
+    const ptrdiff_t nblocks = len / 4;
     const uint8_t* blocks = data + nblocks*4;
 
-    for (int i = -nblocks; i; ++i)
+    for (ptrdiff_t i = -nblocks; i; ++i)
     {
         uint32_t key;
         memcpy(&key, blocks + 4*i, 4);
